fix warnings

+++ 05-SceneManager/WorldMap.cpp
@@ -153,12 +153,12 @@ void CWorldMap::_ParseSection_OBJECTS(string line)
 	{
 		float r = (float)atof(tokens[3].c_str());
 		float b = (float)atof(tokens[4].c_str());
-		int hasPortal = atof(tokens[5].c_str());
-		int scene_id = atof(tokens[6].c_str());
-		int left = atof(tokens[7].c_str());
-		int right = atof(tokens[8].c_str());
-		int up = atof(tokens[9].c_str());
-		int down = atof(tokens[10].c_str());
+		int hasPortal = atoi(tokens[5].c_str());
+		int scene_id = atoi(tokens[6].c_str());
+		int left = atoi(tokens[7].c_str());
+		int right = atoi(tokens[8].c_str());
+		int up = atoi(tokens[9].c_str());
+		int down = atoi(tokens[10].c_str());
 		obj = new CPortal(hasPortal, x,y,r,b,scene_id, left, right, up, down);
 		// General object setup
 		//obj->SetPosition(x, y);
